♻️ Extract `Rule` and `Policy` into their own files
