+++ src/config/mod.rs
@@ -5,14 +5,19 @@ use std::{
 };
 
 use anyhow::{Error, Result};
+use policy::OptionalPolicy;
 use serde::Deserialize;
 use zbus::{Address, AuthMechanism};
 
+pub mod policy;
+pub mod rule;
 mod xml;
 
-use xml::{
-    Document, Element, PolicyContext, PolicyElement, RuleAttributes, RuleElement, TypeElement,
+pub use policy::Policy;
+pub use rule::{
+    Access, ConnectOperation, NameOwnership, Operation, ReceiveOperation, SendOperation,
 };
+use xml::{Document, Element, TypeElement};
 
 /// The bus configuration.
 ///
@@ -171,21 +176,6 @@ impl Config {
 }
 
 #[derive(Clone, Debug, Deserialize, PartialEq)]
-pub struct ConnectOperation {
-    pub group: Option<String>,
-    pub user: Option<String>,
-}
-
-impl From<RuleAttributes> for ConnectOperation {
-    fn from(value: RuleAttributes) -> Self {
-        Self {
-            group: value.group,
-            user: value.user,
-        }
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, PartialEq)]
 #[serde(rename_all = "lowercase")]
 pub enum BusType {
     Session,
@@ -213,356 +203,6 @@ pub enum Name {
     Prefix(String),
 }
 
-#[derive(Clone, Debug, Deserialize, PartialEq)]
-pub enum Operation {
-    /// rules checked when a new connection to the message bus is established
-    Connect(ConnectOperation),
-    /// rules checked when a connection attempts to own a well-known bus names
-    Own(NameOwnership),
-    /// rules that are checked for each recipient of a message
-    Receive(ReceiveOperation),
-    /// rules that are checked when a connection attempts to send a message
-    Send(SendOperation),
-}
-
-type OptionalOperation = Option<Operation>;
-
-impl TryFrom<RuleAttributes> for OptionalOperation {
-    type Error = Error;
-
-    fn try_from(value: RuleAttributes) -> std::result::Result<Self, Self::Error> {
-        let has_connect = value.group.is_some() || value.user.is_some();
-        let has_own = value.own.is_some() || value.own_prefix.is_some();
-        let has_send = value.send_broadcast.is_some()
-            || value.send_destination.is_some()
-            || value.send_destination_prefix.is_some()
-            || value.send_error.is_some()
-            || value.send_interface.is_some()
-            || value.send_member.is_some()
-            || value.send_path.is_some()
-            || value.send_requested_reply.is_some()
-            || value.send_type.is_some();
-        let has_receive = value.receive_error.is_some()
-            || value.receive_interface.is_some()
-            || value.receive_member.is_some()
-            || value.receive_path.is_some()
-            || value.receive_sender.is_some()
-            || value.receive_requested_reply.is_some()
-            || value.receive_type.is_some();
-
-        let operations_count: i8 = vec![has_connect, has_own, has_receive, has_send]
-            .into_iter()
-            .map(i8::from)
-            .sum();
-
-        if operations_count > 1 {
-            return Err(Error::msg(format!("do not mix rule attributes for connect, own, receive, and/or send attributes in the same rule: {value:?}")));
-        }
-
-        if has_connect {
-            Ok(Some(Operation::Connect(ConnectOperation::from(value))))
-        } else if has_own {
-            Ok(Some(Operation::Own(NameOwnership::from(value))))
-        } else if has_receive {
-            Ok(Some(Operation::Receive(ReceiveOperation::from(value))))
-        } else if has_send {
-            Ok(Some(Operation::Send(SendOperation::from(value))))
-        } else {
-            Err(Error::msg(format!("rule must specify supported attributes for connect, own, receive, or send operations: {value:?}")))
-        }
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, PartialEq)]
-pub struct NameOwnership {
-    pub own: Option<Name>,
-}
-
-impl From<RuleAttributes> for NameOwnership {
-    fn from(value: RuleAttributes) -> Self {
-        let own = match value {
-            RuleAttributes {
-                own: Some(some),
-                own_prefix: None,
-                ..
-            } if some == "*" => Some(Name::Any),
-            RuleAttributes {
-                own: Some(some),
-                own_prefix: None,
-                ..
-            } => Some(Name::Exact(some)),
-            RuleAttributes {
-                own: None,
-                own_prefix: Some(some),
-                ..
-            } => Some(Name::Prefix(some)),
-            _ => None,
-        };
-        Self { own }
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, PartialEq)]
-pub enum Policy {
-    DefaultContext(Vec<Rule>),
-    Group(Vec<Rule>, String),
-    MandatoryContext(Vec<Rule>),
-    User(Vec<Rule>, String),
-}
-// TODO: implement Cmp/Ord to help stable-sort Policy values:
-// DefaultContext < Group < User < MandatoryContext
-
-type OptionalPolicy = Option<Policy>;
-
-impl TryFrom<PolicyElement> for OptionalPolicy {
-    type Error = Error;
-
-    fn try_from(value: PolicyElement) -> std::result::Result<Self, Self::Error> {
-        match value {
-            PolicyElement {
-                at_console: Some(_),
-                context: None,
-                group: None,
-                user: None,
-                ..
-            } => Ok(None),
-            PolicyElement {
-                at_console: None,
-                context: Some(c),
-                group: None,
-                rules,
-                user: None,
-            } => Ok(Some(match c {
-                PolicyContext::Default => {
-                    Policy::DefaultContext(rules_try_from_rule_elements(rules)?)
-                }
-                PolicyContext::Mandatory => {
-                    Policy::MandatoryContext(rules_try_from_rule_elements(rules)?)
-                }
-            })),
-            PolicyElement {
-                at_console: None,
-                context: None,
-                group: Some(group),
-                rules,
-                user: None,
-            } => Ok(Some(Policy::Group(
-                rules_try_from_rule_elements(rules)?,
-                group,
-            ))),
-            PolicyElement {
-                at_console: None,
-                context: None,
-                group: None,
-                rules,
-                user: Some(user),
-            } => Ok(Some(Policy::User(
-                rules_try_from_rule_elements(rules)?,
-                user,
-            ))),
-            _ => Err(Error::msg(format!(
-                "policy contains conflicting attributes: {value:?}"
-            ))),
-        }
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, PartialEq)]
-pub struct ReceiveOperation {
-    pub error: Option<String>,
-    pub interface: Option<String>,
-    pub max_fds: Option<u32>,
-    pub member: Option<String>,
-    pub min_fds: Option<u32>,
-    pub path: Option<String>,
-    pub sender: Option<String>,
-    pub r#type: Option<MessageType>,
-}
-
-impl From<RuleAttributes> for ReceiveOperation {
-    fn from(value: RuleAttributes) -> Self {
-        Self {
-            error: value.receive_error,
-            interface: value.receive_interface,
-            max_fds: value.max_fds,
-            member: value.receive_member,
-            min_fds: value.min_fds,
-            path: value.receive_path,
-            sender: value.receive_sender,
-            r#type: value.receive_type,
-        }
-    }
-}
-
-type OptionalRule = Option<Rule>;
-
-impl TryFrom<RuleElement> for OptionalRule {
-    type Error = Error;
-
-    fn try_from(value: RuleElement) -> std::result::Result<Self, Self::Error> {
-        match value {
-            RuleElement::Allow(RuleAttributes {
-                group: Some(_),
-                user: Some(_),
-                ..
-            })
-            | RuleElement::Deny(RuleAttributes {
-                group: Some(_),
-                user: Some(_),
-                ..
-            }) => Err(Error::msg(format!(
-                "`group` cannot be combined with `user` in the same rule: {value:?}"
-            ))),
-            RuleElement::Allow(RuleAttributes {
-                own: Some(_),
-                own_prefix: Some(_),
-                ..
-            })
-            | RuleElement::Deny(RuleAttributes {
-                own: Some(_),
-                own_prefix: Some(_),
-                ..
-            }) => Err(Error::msg(format!(
-                "`own_prefix` cannot be combined with `own` in the same rule: {value:?}"
-            ))),
-            RuleElement::Allow(RuleAttributes {
-                send_destination: Some(_),
-                send_destination_prefix: Some(_),
-                ..
-            })
-            | RuleElement::Deny(RuleAttributes {
-                send_destination: Some(_),
-                send_destination_prefix: Some(_),
-                ..
-            }) => Err(Error::msg(format!(
-                "`send_destination_prefix` cannot be combined with `send_destination` in the same rule: {value:?}"
-            ))),
-            RuleElement::Allow(RuleAttributes {
-                eavesdrop: Some(true),
-                group: None,
-                own: None,
-                receive_requested_reply: None,
-                receive_sender: None,
-                send_broadcast: None,
-                send_destination: None,
-                send_destination_prefix: None,
-                send_error: None,
-                send_interface: None,
-                send_member: None,
-                send_path: None,
-                send_requested_reply: None,
-                send_type: None,
-                user: None,
-                ..
-            }) => {
-                // see: https://github.com/dbus2/busd/pull/146#issuecomment-2408429760
-                Ok(None)
-            }
-            RuleElement::Allow(
-                RuleAttributes {
-                    receive_requested_reply: Some(false),
-                    ..
-                }
-                | RuleAttributes {
-                    send_requested_reply: Some(false),
-                    ..
-                },
-            ) => {
-                // see: https://github.com/dbus2/busd/pull/146#issuecomment-2408429760
-                Ok(None)
-            }
-            RuleElement::Allow(attrs) => {
-                // if attrs.eavesdrop == Some(true) {
-                // see: https://github.com/dbus2/busd/pull/146#issuecomment-2408429760
-                // }
-                match OptionalOperation::try_from(attrs)? {
-                    Some(some) => Ok(Some((Access::Allow, some))),
-                    None => Ok(None),
-                }
-            }
-            RuleElement::Deny(RuleAttributes {
-                eavesdrop: Some(true),
-                ..
-            }) => {
-                // see: https://github.com/dbus2/busd/pull/146#issuecomment-2408429760
-                Ok(None)
-            }
-            RuleElement::Deny(
-                RuleAttributes {
-                    receive_requested_reply: Some(true),
-                    ..
-                }
-                | RuleAttributes {
-                    send_requested_reply: Some(true),
-                    ..
-                },
-            ) => {
-                // see: https://github.com/dbus2/busd/pull/146#issuecomment-2408429760
-                Ok(None)
-            }
-            RuleElement::Deny(attrs) => match OptionalOperation::try_from(attrs)? {
-                Some(some) => Ok(Some((Access::Deny, some))),
-                None => Ok(None),
-            },
-        }
-    }
-}
-
-pub type Rule = (Access, Operation);
-
-#[derive(Clone, Debug, Deserialize, PartialEq)]
-pub enum Access {
-    Allow,
-    Deny,
-}
-
-#[derive(Clone, Debug, Deserialize, PartialEq)]
-pub struct SendOperation {
-    pub broadcast: Option<bool>,
-    pub destination: Option<Name>,
-    pub error: Option<String>,
-    pub interface: Option<String>,
-    pub max_fds: Option<u32>,
-    pub member: Option<String>,
-    pub min_fds: Option<u32>,
-    pub path: Option<String>,
-    pub r#type: Option<MessageType>,
-}
-
-impl From<RuleAttributes> for SendOperation {
-    fn from(value: RuleAttributes) -> Self {
-        let destination = match value {
-            RuleAttributes {
-                send_destination: Some(some),
-                send_destination_prefix: None,
-                ..
-            } if some == "*" => Some(Name::Any),
-            RuleAttributes {
-                send_destination: Some(some),
-                send_destination_prefix: None,
-                ..
-            } => Some(Name::Exact(some)),
-            RuleAttributes {
-                send_destination: None,
-                send_destination_prefix: Some(some),
-                ..
-            } => Some(Name::Prefix(some)),
-            _ => None,
-        };
-        Self {
-            broadcast: value.send_broadcast,
-            destination,
-            error: value.send_error,
-            interface: value.send_interface,
-            max_fds: value.max_fds,
-            member: value.send_member,
-            min_fds: value.min_fds,
-            path: value.send_path,
-            r#type: value.send_type,
-        }
-    }
-}
-
 const DEFAULT_DATA_DIRS: &[&str] = &["/usr/local/share", "/usr/share"];
 
 const STANDARD_SYSTEM_SERVICEDIRS: &[&str] = &[
@@ -570,17 +210,6 @@ const STANDARD_SYSTEM_SERVICEDIRS: &[&str] = &[
     "/usr/share/dbus-1/system-services",
     "/lib/dbus-1/system-services",
 ];
-
-fn rules_try_from_rule_elements(value: Vec<RuleElement>) -> Result<Vec<Rule>> {
-    let mut rules = vec![];
-    for rule in value {
-        let rule = OptionalRule::try_from(rule)?;
-        if let Some(some) = rule {
-            rules.push(some);
-        }
-    }
-    Ok(rules)
-}
 
 fn xdg_data_dirs() -> Vec<PathBuf> {
     if let Ok(ok) = var("XDG_DATA_DIRS") {
@@ -591,6 +220,10 @@ fn xdg_data_dirs() -> Vec<PathBuf> {
 
 #[cfg(test)]
 mod tests {
+    use rule::{
+        Access, ConnectOperation, NameOwnership, Operation, ReceiveOperation, SendOperation,
+    };
+
     use super::*;
 
     #[test]
